requirements.txt
asgiref==3.4.1
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.2.6
django-restframework==0.0.1
djangorestframework==3.12.4
idna==2.9
oauthlib==3.1.0
psycopg2-binary==2.9.1
PyJWT==1.7.1
python-dotenv==0.19.0
python3-openid==3.1.0
pytz==2021.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.9
whitenoise==5.3.0
